% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatNMF.R
\name{concatNMF}
\alias{concatNMF}
\title{Concatenated decomposition with Nonnegative Matrix Factorization}
\usage{
concatNMF(dataset, group, comp_num, weighting = NULL, perturbation = 1e-04)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{perturbation}{the perturbation of the 0 element in the analysis}
}
\value{
A list contains the component and the score of each dataset on every component after concatNMF algorithm
}
\description{
Concatenated decomposition of several matrices with Nonnegative Matrix Factorization (NMF)
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
res_concatNMF = concatNMF(dataset, group, comp_num)

}
\keyword{NMF}
\keyword{pairwise,}
