% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointICA.R
\name{jointICA}
\alias{jointICA}
\title{Joint decomposition with Independent Component Analysis}
\usage{
jointICA()
}
\description{
Joint decomposition of several linked matrices with Independent Component Analysis (ICA)
}
\keyword{ICA}
\keyword{Joint,}
