% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointNMF.R
\name{jointNMF}
\alias{jointNMF}
\title{Joint Decomposition with Nonnegative Matrix Factorization}
\usage{
jointNMF()
}
\description{
Joint decomposition of several linked matrices with Nonnegative Matrix Factorization (NMF)
}
\keyword{NMF}
\keyword{joint,}
