% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseICA.R
\name{pairwiseICA}
\alias{pairwiseICA}
\title{Pairwise decomposition with Independent Component Analysis}
\usage{
pairwiseICA()
}
\description{
Pairwise decomposition of several linked matrices with Independent Component Analysis (ICA)
}
\keyword{ICA}
\keyword{pairwise,}
