% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseNMF.R
\name{pairwiseNMF}
\alias{pairwiseNMF}
\title{Pairwise decomposition with Nonnegative Matrix Factorization}
\usage{
pairwiseNMF()
}
\description{
Pairwise decomposition of several matrices with Nonnegative Matrix Factorization (NMF)
}
\keyword{NMF}
\keyword{pairwise,}
