% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisePCA.R
\name{pairwisePCA}
\alias{pairwisePCA}
\title{Pairwise Decomposition with Principal Component Analysis}
\usage{
pairwisePCA()
}
\description{
Pairwise decomposition of several matrices with Principal Component Analysis (PCA)
}
\keyword{PCA}
\keyword{pairwise,}
