% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{pveMultiple}
\alias{pveMultiple}
\title{Percentage of Variance Explained for Multiple Data sets}
\usage{
pveMultiple(dataset, group, comp_num, list_score, list_component)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{list_score}{A list of extracted scores by the corresponding algorithm}

\item{list_component}{A list of components comptuted by the corresponding algorithm}
}
\value{
The list of scores
}
\description{
Compute the PVE (Percentage of Variance Explained) for multiple data sets on multiple components
}
\keyword{1,}
\keyword{100,}
\keyword{2),}
\keyword{50))}
\keyword{50),}
\keyword{=}
\keyword{Multiple}
\keyword{PVE,}
\keyword{analysis}
\keyword{c(1),}
\keyword{c(1,2),}
\keyword{c(1,3),}
\keyword{c(2),}
\keyword{c(2,2,2,2,2,2,2,2,2)}
\keyword{c(2,4),}
\keyword{c(3),}
\keyword{c(3,4),}
\keyword{c(4))}
\keyword{comp_num}
\keyword{comp_num)}
\keyword{comp_num,}
\keyword{concatPCA(dataset,}
\keyword{dataset}
\keyword{group}
\keyword{group,}
\keyword{list(c(1,2,3,4),}
\keyword{list(matrix(runif(5000,}
\keyword{matrix(runif(5000,}
\keyword{ncol}
\keyword{nrow}
\keyword{pveMultiple(dataset,}
\keyword{res_concatPCA}
\keyword{res_concatPCA$linked_component_list)}
\keyword{res_concatPCA$score_list,}
