% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pve.R
\name{pveSep}
\alias{pveSep}
\title{Percentage of Variance Explained for separate data set}
\usage{
pveSep(dataset, list_score, list_component)
}
\arguments{
\item{dataset}{A list of data sets for input}

\item{list_score}{A list of extracted scores by the corresponding algorithm}

\item{list_component}{A list of components comptuted by the corresponding algorithm}
}
\value{
The list of scores
}
\description{
Compute the PVE (percentage of variance explained) for each data set
}
\keyword{1,}
\keyword{100,}
\keyword{2}
\keyword{2),}
\keyword{50))}
\keyword{=}
\keyword{PVE,}
\keyword{Separate}
\keyword{analysis}
\keyword{comp_num}
\keyword{comp_num)}
\keyword{dataset}
\keyword{list(matrix(runif(5000,}
\keyword{ncol}
\keyword{nrow}
\keyword{pveSep(dataset,}
\keyword{res_sepPCA}
\keyword{res_sepPCA$linked_component_list)}
\keyword{res_sepPCA$score_list,}
\keyword{sepPCA(dataset,}
