% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageiLCA.R
\name{twoStageiLCA}
\alias{twoStageiLCA}
\title{Two-staged Independent Linked Component Analysis}
\usage{
twoStageiLCA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  backup = 0,
  plotting = FALSE
)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{backup}{A positive scalar to determine how many ICs to over select}

\item{plotting}{A boolean value to determine whether to plot the scree plot or not, default to be False}
}
\value{
A list contains the component and the score of each dataset on every component after 2siLCA algorithm
}
\description{
Two-staged Independent Linked Component Analysis, a generalization based on the Two-staged Independent Linked Component Analysis
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1, 2, 3, 4), c(1, 2), c(3, 4), c(1, 3), c(2, 4), c(1), c(2), c(3), c(4))
comp_num = c(2, 2, 2, 2, 2, 2, 2, 2, 2)
res_twoStageiLCA = twoStageiLCA(dataset, group, comp_num)

}
\keyword{LCA}
\keyword{independent}
\keyword{two-staged,}
