% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageLCA.R
\name{twoStageLCA}
\alias{twoStageLCA}
\title{Two-staged Linked Component Analysis}
\usage{
twoStageLCA(dataset, group, comp_num, weighting = NULL, backup = 0)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{weighting}{Weighting of each dataset, initialized to be NULL}

\item{backup}{A positive scalar to determine how many PCs to over select}
}
\value{
A list contains the component and the score of each dataset on every component after 2sLCA algorithm
}
\description{
Two-staged Linked Component Analysis
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1, 2, 3, 4), c(1, 2), c(3, 4), c(1, 3), c(2, 4), c(1), c(2), c(3), c(4))
comp_num = c(2, 2, 2, 2, 2, 2, 2, 2, 2)
res_twoStageLCA = twoStageLCA(dataset, group, comp_num)

}
\keyword{LCA}
\keyword{two-staged,}
