% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointPCA.R
\name{jointPCA}
\alias{jointPCA}
\title{Joint Decomposition with Principal Component Analysis}
\usage{
jointPCA(dataset, group, comp_num, max_ite = 100, max_err = 1e-04)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}

\item{max_ite}{The maximum number of iterations for the jointPCA algorithms to run, default value is set to 100}

\item{max_err}{The maximum error of loss between two iterations, or the program will terminate and return, default value is set to be 0.001}
}
\value{
A list contains the component and the score of each dataset on every component after jointPCA algorithm
}
\description{
Joint decomposition of several linked matrices with Principal Component Analysis (PCA)
}
\keyword{PCA}
\keyword{joint,}
