% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseNMF.R
\name{pairwiseNMF}
\alias{pairwiseNMF}
\title{Pairwise decomposition with Nonnegative Matrix Factorization}
\usage{
pairwiseNMF(dataset, group, comp_num)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}
}
\value{
A list contains the component and the score of each dataset on every component after pairwiseNMF algorithm
}
\description{
Pairwise decomposition of several matrices with Nonnegative Matrix Factorization (NMF)
}
\keyword{NMF}
\keyword{pairwise,}
