% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisePCA.R
\name{pairwisePCA}
\alias{pairwisePCA}
\title{Pairwise Decomposition with Principal Component Analysis}
\usage{
pairwisePCA(dataset, group, comp_num)
}
\arguments{
\item{dataset}{A list of dataset to be analyzed}

\item{group}{A list of grouping of the datasets, indicating the relationship between datasets}

\item{comp_num}{A vector indicates the dimension of each compoent}
}
\value{
A list contains the component and the score of each dataset on every component after pairwisePCA algorithm
}
\description{
Pairwise decomposition of several matrices with Principal Component Analysis (PCA)
}
\examples{
dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1,2,3,4), c(1,2), c(3,4), c(1,3), c(2,4), c(1), c(2), c(3), c(4))
comp_num = c(2,2,2,2,2,2,2,2,2)
res_pairwisePCA = pairwisePCA(dataset, group, comp_num)

}
\keyword{PCA}
\keyword{pairwise,}
